set-up heredoc signal interrupt handling, edited  a few of Josh's functions but nothing material for it to impact the overall code handling

+++ freedom.c
@@ -6,7 +6,7 @@
 /*   By: esezalor <esezalor@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/09 14:10:43 by esezalor          #+#    #+#             */
-/*   Updated: 2026/04/06 12:28:51 by esezalor         ###   ########.fr       */
+/*   Updated: 2026/04/06 15:11:47 by esezalor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,10 @@ void	unlink_files(t_cmd *cmds)
 	while (current)
 	{
 		if (current->infile && current->heredoc_delim)
-			unlink(current->infile);
+		{
+			if(access(current->infile, F_OK) != -1)
+				unlink(current->infile);
+		}
 		current = current->next;
 	}
 }
